+++ CHANGELOG.md
@@ -15,8 +15,8 @@ where something in said state is leading to an uncaught exception.
 - [Web] For operating systems with support for selecting a preferred light /dark
 theme, the application now respects the system preference by default.
 - sensuctl dump can now list the types of supported resources with --types.
-- Added the `Version` field to the `Entity` resource, which reflects the
-Sensu semver version of the agent entity.
+- Added the `sensu_agent_version` field to the `Entity` resource, which reflects
+the Sensu semver version of the agent entity.
 
 ### Changed
 - [Web] Github is not always the best place for feature requests and discussion,